Refactored customImage to use new Next image component
