+++ components/blocks/customImage.tsx
@@ -1,8 +1,7 @@
 import React from "react";
 import type { Template } from "tinacms";
-import Image from "next/legacy/image";
+import Image from "next/image";
 
-/* TODO: refactor with next/image */
 export const CustomImage = ({ data }) => {
   return (
     <Image
@@ -10,6 +9,7 @@ export const CustomImage = ({ data }) => {
       alt={data.altText}
       height={data.height}
       width={data.width}
+      className="inline-block"
     />
   );
 };
