+++ content/pages/home.mdx
@@ -154,6 +154,6 @@ SSW Consulting has over 30 years of experience developing awesome Microsoft solu
 
 ---
 
-### Bring your apps to China <CustomImage src="/images/logos/china-flag.svg" altText="Chinese Flag" height={25} width={30} />
+### Bring your apps to China <Flag country="China" />
 
 China is a booming market and now is the time to take advantage of this growing user base. If you have a successful application that you would like to bring to the Chinese market, then working with [SSW can help](/consulting/chinafy-app) streamline your entry into this market.
